Add tag extraction feature to log_tagger module
Each tags config may now contain a 'extractors' field. In that
field a list of regular expressions may be stored which will be
used to extract a dynamic tag value from the source field.
The extraction is done via groups and therefore those expressions
need to at least have a single group. In case there is more than
one group in an expression only the first one will be used.